IN PROGRESS

+++ single_patient_test/Provider.h
@@ -7,7 +7,7 @@ protected:
 public:
 
 	Provider(Patient*);
-	void take_vital_signs(int, int, int, int, float, int, int);
+	void take_vital_signs(int, int, int, int, double, int, int);
 	void chart_report();
 	void pause_continue();
