+++ single_patient_test/vitals.cpp
@@ -3,7 +3,7 @@
 
 //class Vitals definitions
 
-Vitals::Vitals(int sys, int dia, int hr, int rr, float t, int ox, int lc, std::string clock) //systolic, diastolic, heart rate, resp, temp, level conciousness
+Vitals::Vitals(int sys, int dia, int hr, int rr, double t, int ox, int lc, std::string clock) //systolic, diastolic, heart rate, resp, temp, level conciousness
 	:systolic(sys), diastolic(dia), pulse(hr), resps(rr), temp(t), o2_sat(ox), loc(lc), time_taken(clock)
 {
 
@@ -26,7 +26,7 @@ int Vitals::show_diastolic()
 	return Vitals::diastolic;
 }
 
-float Vitals::show_temp()
+double Vitals::show_temp()
 {
 	return Vitals::temp;
 }
